bug fix: Can only calculate the mean of floating types. Got Long instead.

+++ federated-learning/utils/util.py
@@ -313,7 +313,7 @@ def extract_sign_by_diff(w_local, w_glob, momentum, beta):
         momentum[k] = torch.add(torch.mul(momentum[k], beta), torch.mul(sgd[k], 1-beta))
         # w_signed[k] = torch.sign(momentum[k])
         w_signed[k] = torch.sign(sgd[k])
-        sgd_k_mean = torch.mean(sgd[k]).item()
+        sgd_k_mean = torch.mean(sgd[k].type(torch.FloatTensor)).item()
         try:
             precision = math.floor(math.log10(abs(sgd_k_mean)))
         except ValueError:
